refactor(core): add global-shortcut Cargo feature, enhancing binary size (#3956)
