+++ core/tauri-runtime-wry/Cargo.toml
@@ -41,3 +41,4 @@ macos-private-api = [
 objc-exception = [ "wry/objc-exception" ]
 gtk-tray = [ "wry/gtk-tray" ]
 ayatana-tray = [ "wry/ayatana-tray" ]
+global-shortcut = [ "tauri-runtime/global-shortcut" ]
