+++ core/tauri-runtime/Cargo.toml
@@ -46,3 +46,4 @@ gtk = { version = "0.15", features = [ "v3_20" ] }
 devtools = [ ]
 system-tray = [ ]
 macos-private-api = [ ]
+global-shortcut = [ ]
