+++ core/tauri/src/test/mock_runtime.rs
@@ -12,8 +12,8 @@ use tauri_runtime::{
     dpi::{PhysicalPosition, PhysicalSize, Position, Size},
     CursorIcon, DetachedWindow, MenuEvent, PendingWindow, WindowEvent,
   },
-  ClipboardManager, Dispatch, EventLoopProxy, GlobalShortcutManager, Result, RunEvent, Runtime,
-  RuntimeHandle, UserAttentionType, UserEvent, WindowIcon,
+  ClipboardManager, Dispatch, EventLoopProxy, Result, RunEvent, Runtime, RuntimeHandle,
+  UserAttentionType, UserEvent, WindowIcon,
 };
 #[cfg(feature = "system-tray")]
 use tauri_runtime::{
@@ -92,12 +92,14 @@ pub struct MockDispatcher {
   context: RuntimeContext,
 }
 
+#[cfg(feature = "global-shortcut")]
 #[derive(Debug, Clone)]
 pub struct MockGlobalShortcutManager {
   context: RuntimeContext,
 }
 
-impl GlobalShortcutManager for MockGlobalShortcutManager {
+#[cfg(feature = "global-shortcut")]
+impl tauri_runtime::GlobalShortcutManager for MockGlobalShortcutManager {
   fn is_registered(&self, accelerator: &str) -> Result<bool> {
     Ok(
       self
@@ -543,6 +545,7 @@ impl<T: UserEvent> EventLoopProxy<T> for EventProxy {
 #[derive(Debug)]
 pub struct MockRuntime {
   pub context: RuntimeContext,
+  #[cfg(feature = "global-shortcut")]
   global_shortcut_manager: MockGlobalShortcutManager,
   clipboard_manager: MockClipboardManager,
   #[cfg(feature = "system-tray")]
@@ -556,6 +559,7 @@ impl MockRuntime {
       clipboard: Default::default(),
     };
     Self {
+      #[cfg(feature = "global-shortcut")]
       global_shortcut_manager: MockGlobalShortcutManager {
         context: context.clone(),
       },
@@ -574,6 +578,7 @@ impl MockRuntime {
 impl<T: UserEvent> Runtime<T> for MockRuntime {
   type Dispatcher = MockDispatcher;
   type Handle = MockRuntimeHandle;
+  #[cfg(feature = "global-shortcut")]
   type GlobalShortcutManager = MockGlobalShortcutManager;
   type ClipboardManager = MockClipboardManager;
   #[cfg(feature = "system-tray")]
@@ -599,6 +604,7 @@ impl<T: UserEvent> Runtime<T> for MockRuntime {
     }
   }
 
+  #[cfg(feature = "global-shortcut")]
   fn global_shortcut_manager(&self) -> Self::GlobalShortcutManager {
     self.global_shortcut_manager.clone()
   }
